add test for PokemonSpecies component
