+++ client/src/mocks/pokemon-specie.js
@@ -1,0 +1,906 @@
+const pokemonSpecieData = {
+  base_happiness: 70,
+  capture_rate: 45,
+  color: {
+    name: 'green',
+    url: 'https://pokeapi.co/api/v2/pokemon-color/5/',
+  },
+  egg_groups: [
+    {
+      name: 'plant',
+      url: 'https://pokeapi.co/api/v2/egg-group/7/',
+    },
+    {
+      name: 'monster',
+      url: 'https://pokeapi.co/api/v2/egg-group/1/',
+    },
+  ],
+  evolution_chain: {
+    url: 'https://pokeapi.co/api/v2/evolution-chain/1/',
+  },
+  evolves_from_species: null,
+  flavor_text_entries: [
+    {
+      flavor_text:
+        '日なたで　昼寝を　する　姿を　見かける。\n太陽の　光を　いっぱい　浴びることで\n背中の　タネが　大きく　育つのだ。',
+      language: {
+        name: 'ja',
+        url: 'https://pokeapi.co/api/v2/language/11/',
+      },
+      version: {
+        name: 'alpha-sapphire',
+        url: 'https://pokeapi.co/api/v2/version/26/',
+      },
+    },
+    {
+      flavor_text:
+        'Bulbasaur can be seen napping in bright sunlight.\nThere is a seed on its back. By soaking up the sun’s rays,\nthe seed grows progressively larger.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'alpha-sapphire',
+        url: 'https://pokeapi.co/api/v2/version/26/',
+      },
+    },
+    {
+      flavor_text:
+        'È possibile vedere Bulbasaur mentre schiaccia un pisolino\nsotto il sole. Ha un seme piantato sulla schiena. Grazie ai\nraggi solari il seme cresce ingrandendosi progressivamente.',
+      language: {
+        name: 'it',
+        url: 'https://pokeapi.co/api/v2/language/8/',
+      },
+      version: {
+        name: 'alpha-sapphire',
+        url: 'https://pokeapi.co/api/v2/version/26/',
+      },
+    },
+    {
+      flavor_text:
+        'A Bulbasaur es fácil verle echándose una siesta al sol.\nLa semilla que tiene en el lomo va creciendo cada vez más\na medida que absorbe los rayos del sol.',
+      language: {
+        name: 'es',
+        url: 'https://pokeapi.co/api/v2/language/7/',
+      },
+      version: {
+        name: 'alpha-sapphire',
+        url: 'https://pokeapi.co/api/v2/version/26/',
+      },
+    },
+    {
+      flavor_text:
+        'Bisasam macht gern einmal ein Nickerchen im\nSonnenschein. Auf seinem Rücken trägt es einen\nSamen. Indem es Sonnenstrahlen aufsaugt,\nwird er zunehmend größer.',
+      language: {
+        name: 'de',
+        url: 'https://pokeapi.co/api/v2/language/6/',
+      },
+      version: {
+        name: 'alpha-sapphire',
+        url: 'https://pokeapi.co/api/v2/version/26/',
+      },
+    },
+    {
+      flavor_text:
+        'Bulbizarre passe son temps à faire la sieste sous le soleil.\nIl y a une graine sur son dos. Il absorbe les rayons du soleil\npour faire doucement pousser la graine.',
+      language: {
+        name: 'fr',
+        url: 'https://pokeapi.co/api/v2/language/5/',
+      },
+      version: {
+        name: 'alpha-sapphire',
+        url: 'https://pokeapi.co/api/v2/version/26/',
+      },
+    },
+    {
+      flavor_text:
+        '양지에서 낮잠 자는 모습을 볼 수 있다.\n태양의 빛을 많이 받으면\n등의 씨앗이 크게 자란다.',
+      language: {
+        name: 'ko',
+        url: 'https://pokeapi.co/api/v2/language/3/',
+      },
+      version: {
+        name: 'alpha-sapphire',
+        url: 'https://pokeapi.co/api/v2/version/26/',
+      },
+    },
+    {
+      flavor_text:
+        'ひなたで　ひるねを　する　すがたを　\u202fかける。\nたいようの　ひかりを　いっぱい　あびることで\nせなかの　タネが　おおきく　そだつのだ。',
+      language: {
+        name: 'ja-Hrkt',
+        url: 'https://pokeapi.co/api/v2/language/1/',
+      },
+      version: {
+        name: 'alpha-sapphire',
+        url: 'https://pokeapi.co/api/v2/version/26/',
+      },
+    },
+    {
+      flavor_text:
+        '日なたで　昼寝を　する　姿を　見かける。\n太陽の　光を　いっぱい　浴びることで\n背中の　タネが　大きく　育つのだ。',
+      language: {
+        name: 'ja',
+        url: 'https://pokeapi.co/api/v2/language/11/',
+      },
+      version: {
+        name: 'omega-ruby',
+        url: 'https://pokeapi.co/api/v2/version/25/',
+      },
+    },
+    {
+      flavor_text:
+        'Bulbasaur can be seen napping in bright sunlight.\nThere is a seed on its back. By soaking up the sun’s rays,\nthe seed grows progressively larger.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'omega-ruby',
+        url: 'https://pokeapi.co/api/v2/version/25/',
+      },
+    },
+    {
+      flavor_text:
+        'È possibile vedere Bulbasaur mentre schiaccia un pisolino\nsotto il sole. Ha un seme piantato sulla schiena. Grazie ai\nraggi solari il seme cresce ingrandendosi progressivamente.',
+      language: {
+        name: 'it',
+        url: 'https://pokeapi.co/api/v2/language/8/',
+      },
+      version: {
+        name: 'omega-ruby',
+        url: 'https://pokeapi.co/api/v2/version/25/',
+      },
+    },
+    {
+      flavor_text:
+        'A Bulbasaur es fácil verle echándose una siesta al sol.\nLa semilla que tiene en el lomo va creciendo cada vez más\na medida que absorbe los rayos del sol.',
+      language: {
+        name: 'es',
+        url: 'https://pokeapi.co/api/v2/language/7/',
+      },
+      version: {
+        name: 'omega-ruby',
+        url: 'https://pokeapi.co/api/v2/version/25/',
+      },
+    },
+    {
+      flavor_text:
+        'Bisasam macht gern einmal ein Nickerchen im\nSonnenschein. Auf seinem Rücken trägt es einen\nSamen. Indem es Sonnenstrahlen aufsaugt,\nwird der Samen zunehmend größer.',
+      language: {
+        name: 'de',
+        url: 'https://pokeapi.co/api/v2/language/6/',
+      },
+      version: {
+        name: 'omega-ruby',
+        url: 'https://pokeapi.co/api/v2/version/25/',
+      },
+    },
+    {
+      flavor_text:
+        'Bulbizarre passe son temps à faire la sieste sous le soleil.\nIl y a une graine sur son dos. Il absorbe les rayons du soleil\npour faire doucement pousser la graine.',
+      language: {
+        name: 'fr',
+        url: 'https://pokeapi.co/api/v2/language/5/',
+      },
+      version: {
+        name: 'omega-ruby',
+        url: 'https://pokeapi.co/api/v2/version/25/',
+      },
+    },
+    {
+      flavor_text:
+        '양지에서 낮잠 자는 모습을 볼 수 있다.\n태양의 빛을 많이 받으면\n등의 씨앗이 크게 자란다.',
+      language: {
+        name: 'ko',
+        url: 'https://pokeapi.co/api/v2/language/3/',
+      },
+      version: {
+        name: 'omega-ruby',
+        url: 'https://pokeapi.co/api/v2/version/25/',
+      },
+    },
+    {
+      flavor_text:
+        'ひなたで　ひるねを　する　すがたを　\u202fかける。\nたいようの　ひかりを　いっぱい　あびることで\nせなかの　タネが　おおきく　そだつのだ。',
+      language: {
+        name: 'ja-Hrkt',
+        url: 'https://pokeapi.co/api/v2/language/1/',
+      },
+      version: {
+        name: 'omega-ruby',
+        url: 'https://pokeapi.co/api/v2/version/25/',
+      },
+    },
+    {
+      flavor_text:
+        '生まれてから　しばらくの　あいだは\n背中の　タネから　栄養を　もらって\n大きく　育つ。',
+      language: {
+        name: 'ja',
+        url: 'https://pokeapi.co/api/v2/language/11/',
+      },
+      version: {
+        name: 'y',
+        url: 'https://pokeapi.co/api/v2/version/24/',
+      },
+    },
+    {
+      flavor_text:
+        'For some time after its birth, it grows by gaining\nnourishment from the seed on its back.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'y',
+        url: 'https://pokeapi.co/api/v2/version/24/',
+      },
+    },
+    {
+      flavor_text:
+        'Dopo la nascita, cresce traendo nutrimento dal seme\npiantato sul suo dorso.',
+      language: {
+        name: 'it',
+        url: 'https://pokeapi.co/api/v2/language/8/',
+      },
+      version: {
+        name: 'y',
+        url: 'https://pokeapi.co/api/v2/version/24/',
+      },
+    },
+    {
+      flavor_text:
+        'Después de nacer, crece alimentándose de las\nsemillas de su lomo.',
+      language: {
+        name: 'es',
+        url: 'https://pokeapi.co/api/v2/language/7/',
+      },
+      version: {
+        name: 'y',
+        url: 'https://pokeapi.co/api/v2/version/24/',
+      },
+    },
+    {
+      flavor_text:
+        'Nach der Geburt nimmt es für eine Weile Nährstoffe\nüber den Samen auf seinem Rücken auf.',
+      language: {
+        name: 'de',
+        url: 'https://pokeapi.co/api/v2/language/6/',
+      },
+      version: {
+        name: 'y',
+        url: 'https://pokeapi.co/api/v2/version/24/',
+      },
+    },
+    {
+      flavor_text:
+        'Au matin de sa vie, la graine sur son dos lui fournit\nles éléments dont il a besoin pour grandir.',
+      language: {
+        name: 'fr',
+        url: 'https://pokeapi.co/api/v2/language/5/',
+      },
+      version: {
+        name: 'y',
+        url: 'https://pokeapi.co/api/v2/version/24/',
+      },
+    },
+    {
+      flavor_text:
+        '태어나서부터 얼마 동안은\n등의 씨앗으로부터 영양을\n공급받아 크게 성장한다.',
+      language: {
+        name: 'ko',
+        url: 'https://pokeapi.co/api/v2/language/3/',
+      },
+      version: {
+        name: 'y',
+        url: 'https://pokeapi.co/api/v2/version/24/',
+      },
+    },
+    {
+      flavor_text:
+        'うまれてから　しばらくの　あいだは\nせなかの　タネから　えいようを\nもらって　おおきく　そだつ。',
+      language: {
+        name: 'ja-Hrkt',
+        url: 'https://pokeapi.co/api/v2/language/1/',
+      },
+      version: {
+        name: 'y',
+        url: 'https://pokeapi.co/api/v2/version/24/',
+      },
+    },
+    {
+      flavor_text:
+        '生まれたときから　背中に\n不思議な　タネが　植えてあって\n体と　ともに　育つという。',
+      language: {
+        name: 'ja',
+        url: 'https://pokeapi.co/api/v2/language/11/',
+      },
+      version: {
+        name: 'x',
+        url: 'https://pokeapi.co/api/v2/version/23/',
+      },
+    },
+    {
+      flavor_text:
+        'A strange seed was planted on its back at birth.\nThe plant sprouts and grows with this Pokémon.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'x',
+        url: 'https://pokeapi.co/api/v2/version/23/',
+      },
+    },
+    {
+      flavor_text:
+        'Alla nascita gli è stato piantato sulla schiena un seme\nraro. La pianta sboccia e cresce con lui.',
+      language: {
+        name: 'it',
+        url: 'https://pokeapi.co/api/v2/language/8/',
+      },
+      version: {
+        name: 'x',
+        url: 'https://pokeapi.co/api/v2/version/23/',
+      },
+    },
+    {
+      flavor_text:
+        'Una rara semilla le fue plantada en el lomo al nacer.\nLa planta brota y crece con este Pokémon.',
+      language: {
+        name: 'es',
+        url: 'https://pokeapi.co/api/v2/language/7/',
+      },
+      version: {
+        name: 'x',
+        url: 'https://pokeapi.co/api/v2/version/23/',
+      },
+    },
+    {
+      flavor_text:
+        'Dieses Pokémon trägt von Geburt an einen Samen\nauf dem Rücken, der mit ihm keimt und wächst.',
+      language: {
+        name: 'de',
+        url: 'https://pokeapi.co/api/v2/language/6/',
+      },
+      version: {
+        name: 'x',
+        url: 'https://pokeapi.co/api/v2/version/23/',
+      },
+    },
+    {
+      flavor_text:
+        'Il a une étrange graine plantée sur son dos.\nElle grandit avec lui depuis sa naissance.',
+      language: {
+        name: 'fr',
+        url: 'https://pokeapi.co/api/v2/language/5/',
+      },
+      version: {
+        name: 'x',
+        url: 'https://pokeapi.co/api/v2/version/23/',
+      },
+    },
+    {
+      flavor_text:
+        '태어났을 때부터 등에\n이상한 씨앗이 심어져 있으며\n몸과 함께 자란다고 한다.',
+      language: {
+        name: 'ko',
+        url: 'https://pokeapi.co/api/v2/language/3/',
+      },
+      version: {
+        name: 'x',
+        url: 'https://pokeapi.co/api/v2/version/23/',
+      },
+    },
+    {
+      flavor_text:
+        'うまれたときから　せなかに\nふしぎな　タネが　うえてあって\nからだと　ともに　そだつという。',
+      language: {
+        name: 'ja-Hrkt',
+        url: 'https://pokeapi.co/api/v2/language/1/',
+      },
+      version: {
+        name: 'x',
+        url: 'https://pokeapi.co/api/v2/version/23/',
+      },
+    },
+    {
+      flavor_text:
+        'For some time after its birth, it\ngrows by gaining nourishment from\nthe seed on its back.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'white-2',
+        url: 'https://pokeapi.co/api/v2/version/22/',
+      },
+    },
+    {
+      flavor_text:
+        'For some time after its birth, it\ngrows by gaining nourishment from\nthe seed on its back.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'black-2',
+        url: 'https://pokeapi.co/api/v2/version/21/',
+      },
+    },
+    {
+      flavor_text:
+        'For some time after its birth, it\ngrows by gaining nourishment from\nthe seed on its back.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'white',
+        url: 'https://pokeapi.co/api/v2/version/18/',
+      },
+    },
+    {
+      flavor_text:
+        'Au matin de sa vie, la graine sur\nson dos lui fournit les éléments\ndont il a besoin pour grandir.',
+      language: {
+        name: 'fr',
+        url: 'https://pokeapi.co/api/v2/language/5/',
+      },
+      version: {
+        name: 'white',
+        url: 'https://pokeapi.co/api/v2/version/18/',
+      },
+    },
+    {
+      flavor_text:
+        'For some time after its birth, it\ngrows by gaining nourishment from\nthe seed on its back.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'black',
+        url: 'https://pokeapi.co/api/v2/version/17/',
+      },
+    },
+    {
+      flavor_text:
+        'Au matin de sa vie, la graine sur\nson dos lui fournit les éléments\ndont il a besoin pour grandir.',
+      language: {
+        name: 'fr',
+        url: 'https://pokeapi.co/api/v2/language/5/',
+      },
+      version: {
+        name: 'black',
+        url: 'https://pokeapi.co/api/v2/version/17/',
+      },
+    },
+    {
+      flavor_text:
+        'It carries a seed on its back right\nfrom birth. As it grows older, the\nseed also grows larger.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'soulsilver',
+        url: 'https://pokeapi.co/api/v2/version/16/',
+      },
+    },
+    {
+      flavor_text:
+        'The seed on its back is filled\nwith nutrients. The seed grows\nsteadily larger as its body grows.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'heartgold',
+        url: 'https://pokeapi.co/api/v2/version/15/',
+      },
+    },
+    {
+      flavor_text:
+        'For some time after its birth, it\ngrows by gaining nourishment from\nthe seed on its back.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'platinum',
+        url: 'https://pokeapi.co/api/v2/version/14/',
+      },
+    },
+    {
+      flavor_text:
+        'For some time after its birth, it\ngrows by gaining nourishment from\nthe seed on its back.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'pearl',
+        url: 'https://pokeapi.co/api/v2/version/13/',
+      },
+    },
+    {
+      flavor_text:
+        'For some time after its birth, it\ngrows by gaining nourishment from\nthe seed on its back.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'diamond',
+        url: 'https://pokeapi.co/api/v2/version/12/',
+      },
+    },
+    {
+      flavor_text:
+        'A strange seed was planted on its back at\nbirth. The plant sprouts and grows with\nthis POKéMON.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'leafgreen',
+        url: 'https://pokeapi.co/api/v2/version/11/',
+      },
+    },
+    {
+      flavor_text:
+        'There is a plant seed on its back right\nfrom the day this POKéMON is born.\nThe seed slowly grows larger.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'firered',
+        url: 'https://pokeapi.co/api/v2/version/10/',
+      },
+    },
+    {
+      flavor_text:
+        'BULBASAUR can be seen napping in bright\nsunlight. There is a seed on its back.\nBy soaking up the sun’s rays, the seed\ngrows progressively larger.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'emerald',
+        url: 'https://pokeapi.co/api/v2/version/9/',
+      },
+    },
+    {
+      flavor_text:
+        'BULBASAUR can be seen napping in\nbright sunlight.\nThere is a seed on its back.\fBy soaking up the sun’s rays, the seed\ngrows progressively larger.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'sapphire',
+        url: 'https://pokeapi.co/api/v2/version/8/',
+      },
+    },
+    {
+      flavor_text:
+        'BULBASAUR can be seen napping in\nbright sunlight.\nThere is a seed on its back.\fBy soaking up the sun’s rays, the seed\ngrows progressively larger.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'ruby',
+        url: 'https://pokeapi.co/api/v2/version/7/',
+      },
+    },
+    {
+      flavor_text:
+        'While it is young,\nit uses the\nnutrients that are\fstored in the\nseeds on its back\nin order to grow.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'crystal',
+        url: 'https://pokeapi.co/api/v2/version/6/',
+      },
+    },
+    {
+      flavor_text:
+        'It carries a seed\non its back right\nfrom birth. As it\fgrows older, the\nseed also grows\nlarger.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'silver',
+        url: 'https://pokeapi.co/api/v2/version/5/',
+      },
+    },
+    {
+      flavor_text:
+        'The seed on its\nback is filled\nwith nutrients.\fThe seed grows\nsteadily larger as\nits body grows.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'gold',
+        url: 'https://pokeapi.co/api/v2/version/4/',
+      },
+    },
+    {
+      flavor_text:
+        'It can go for days\nwithout eating a\nsingle morsel.\fIn the bulb on\nits back, it\nstores energy.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'yellow',
+        url: 'https://pokeapi.co/api/v2/version/3/',
+      },
+    },
+    {
+      flavor_text:
+        'A strange seed was\nplanted on its\nback at birth.\fThe plant sprouts\nand grows with\nthis POKéMON.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'blue',
+        url: 'https://pokeapi.co/api/v2/version/2/',
+      },
+    },
+    {
+      flavor_text:
+        'A strange seed was\nplanted on its\nback at birth.\fThe plant sprouts\nand grows with\nthis POKéMON.',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      version: {
+        name: 'red',
+        url: 'https://pokeapi.co/api/v2/version/1/',
+      },
+    },
+  ],
+  form_descriptions: [],
+  forms_switchable: false,
+  gender_rate: 1,
+  genera: [
+    {
+      genus: '种子宝可梦',
+      language: {
+        name: 'zh-Hans',
+        url: 'https://pokeapi.co/api/v2/language/12/',
+      },
+    },
+    {
+      genus: 'たねポケモン',
+      language: {
+        name: 'ja',
+        url: 'https://pokeapi.co/api/v2/language/11/',
+      },
+    },
+    {
+      genus: 'Seed Pokémon',
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+    },
+    {
+      genus: 'Pokémon Seme',
+      language: {
+        name: 'it',
+        url: 'https://pokeapi.co/api/v2/language/8/',
+      },
+    },
+    {
+      genus: 'Pokémon Semilla',
+      language: {
+        name: 'es',
+        url: 'https://pokeapi.co/api/v2/language/7/',
+      },
+    },
+    {
+      genus: 'Samen',
+      language: {
+        name: 'de',
+        url: 'https://pokeapi.co/api/v2/language/6/',
+      },
+    },
+    {
+      genus: 'Pokémon Graine',
+      language: {
+        name: 'fr',
+        url: 'https://pokeapi.co/api/v2/language/5/',
+      },
+    },
+    {
+      genus: '種子寶可夢',
+      language: {
+        name: 'zh-Hant',
+        url: 'https://pokeapi.co/api/v2/language/4/',
+      },
+    },
+    {
+      genus: '씨앗포켓몬',
+      language: {
+        name: 'ko',
+        url: 'https://pokeapi.co/api/v2/language/3/',
+      },
+    },
+    {
+      genus: 'たねポケモン',
+      language: {
+        name: 'ja-Hrkt',
+        url: 'https://pokeapi.co/api/v2/language/1/',
+      },
+    },
+  ],
+  generation: {
+    name: 'generation-i',
+    url: 'https://pokeapi.co/api/v2/generation/1/',
+  },
+  growth_rate: {
+    name: 'medium-slow',
+    url: 'https://pokeapi.co/api/v2/growth-rate/4/',
+  },
+  habitat: {
+    name: 'grassland',
+    url: 'https://pokeapi.co/api/v2/pokemon-habitat/3/',
+  },
+  has_gender_differences: false,
+  hatch_counter: 20,
+  id: 1,
+  is_baby: false,
+  name: 'bulbasaur',
+  names: [
+    {
+      language: {
+        name: 'zh-Hans',
+        url: 'https://pokeapi.co/api/v2/language/12/',
+      },
+      name: '妙蛙种子',
+    },
+    {
+      language: {
+        name: 'ja',
+        url: 'https://pokeapi.co/api/v2/language/11/',
+      },
+      name: 'フシギダネ',
+    },
+    {
+      language: {
+        name: 'en',
+        url: 'https://pokeapi.co/api/v2/language/9/',
+      },
+      name: 'Bulbasaur',
+    },
+    {
+      language: {
+        name: 'it',
+        url: 'https://pokeapi.co/api/v2/language/8/',
+      },
+      name: 'Bulbasaur',
+    },
+    {
+      language: {
+        name: 'es',
+        url: 'https://pokeapi.co/api/v2/language/7/',
+      },
+      name: 'Bulbasaur',
+    },
+    {
+      language: {
+        name: 'de',
+        url: 'https://pokeapi.co/api/v2/language/6/',
+      },
+      name: 'Bisasam',
+    },
+    {
+      language: {
+        name: 'fr',
+        url: 'https://pokeapi.co/api/v2/language/5/',
+      },
+      name: 'Bulbizarre',
+    },
+    {
+      language: {
+        name: 'zh-Hant',
+        url: 'https://pokeapi.co/api/v2/language/4/',
+      },
+      name: '妙蛙種子',
+    },
+    {
+      language: {
+        name: 'ko',
+        url: 'https://pokeapi.co/api/v2/language/3/',
+      },
+      name: '이상해씨',
+    },
+    {
+      language: {
+        name: 'roomaji',
+        url: 'https://pokeapi.co/api/v2/language/2/',
+      },
+      name: 'Fushigidane',
+    },
+    {
+      language: {
+        name: 'ja-Hrkt',
+        url: 'https://pokeapi.co/api/v2/language/1/',
+      },
+      name: 'フシギダネ',
+    },
+  ],
+  order: 1,
+  pal_park_encounters: [
+    {
+      area: {
+        name: 'field',
+        url: 'https://pokeapi.co/api/v2/pal-park-area/2/',
+      },
+      base_score: 50,
+      rate: 30,
+    },
+  ],
+  pokedex_numbers: [
+    {
+      entry_number: 80,
+      pokedex: {
+        name: 'kalos-central',
+        url: 'https://pokeapi.co/api/v2/pokedex/12/',
+      },
+    },
+    {
+      entry_number: 231,
+      pokedex: {
+        name: 'updated-johto',
+        url: 'https://pokeapi.co/api/v2/pokedex/7/',
+      },
+    },
+    {
+      entry_number: 226,
+      pokedex: {
+        name: 'original-johto',
+        url: 'https://pokeapi.co/api/v2/pokedex/3/',
+      },
+    },
+    {
+      entry_number: 1,
+      pokedex: {
+        name: 'kanto',
+        url: 'https://pokeapi.co/api/v2/pokedex/2/',
+      },
+    },
+    {
+      entry_number: 1,
+      pokedex: {
+        name: 'national',
+        url: 'https://pokeapi.co/api/v2/pokedex/1/',
+      },
+    },
+  ],
+  shape: {
+    name: 'quadruped',
+    url: 'https://pokeapi.co/api/v2/pokemon-shape/8/',
+  },
+  varieties: [
+    {
+      is_default: true,
+      pokemon: {
+        name: 'bulbasaur',
+        url: 'https://pokeapi.co/api/v2/pokemon/1/',
+      },
+    },
+  ],
+};
+
+export default pokemonSpecieData;
